extract readonly tests, add more tests
